Alright moved everything over, need to add tests though and I don't feel like doing that tonight haha.
